start cleaning up some of the utils and write some basic stuff that will load mscoco
