+++ train_capsnet.py
@@ -23,8 +23,6 @@ logger = daiquiri.getLogger(__name__)
 def main(args):
     """Get dataset hyperparameters."""
     assert len(args) == 2 and isinstance(args[1], str)
-    dataset_name = args[1]
-    logger.info('Using dataset: {}'.format(dataset_name))
 
     """Set reproduciable random seed"""
     tf.set_random_seed(1234)
@@ -32,7 +30,7 @@ def main(args):
     coord_add = get_coord_add(dataset_name)
     dataset_size = get_dataset_size_train(dataset_name)
     num_classes = get_num_classes(dataset_name)
-    create_inputs = get_create_inputs(dataset_name, is_train=True, epochs=cfg.epoch)
+    create_inputs = get_create_inputs(dataset_name, cfg, is_train=True, epochs=cfg.epoch)
 
     with tf.Graph().as_default(), tf.device('/cpu:0'):
         """Get global_step."""
